+++ pkg/ruler/compat.go
@@ -177,7 +177,20 @@ func EngineQueryFunc(engine promql.QueryEngine, frontendClient *frontendClient, 
 		}
 
 		if frontendClient != nil {
-			v, err := frontendClient.InstantQuery(ctx, qs, t)
+			// query parameters sent to the Query Frontend to leave rule information logs on query stats
+			queryParams := map[string]string{}
+
+			if origin := ctx.Value(promql.QueryOrigin{}); origin != nil {
+				queryLabels := origin.(map[string]interface{})
+				rgMap := queryLabels["ruleGroup"].(map[string]string)
+				queryParams["rule_group"] = rgMap["name"]
+				queryParams["rule_namespace"] = rgMap["file"]
+			}
+			ruleDetail := rules.FromOriginContext(ctx)
+			queryParams["rule"] = ruleDetail.Name
+			queryParams["rule_kind"] = ruleDetail.Kind
+
+			v, err := frontendClient.InstantQuery(ctx, qs, t, queryParams)
 			if err != nil {
 				return nil, err
 			}